requirements.txt
dependency_injector
pre-commit
py-automapper
pydantic
pyodbc
pytz
sqlalchemy
